+++ angular/projects/wcs-angular/src/lib/proxies.ts
@@ -1444,7 +1444,7 @@ export class WcsSwitch {
   constructor(c: ChangeDetectorRef, r: ElementRef, protected z: NgZone) {
     c.detach();
     this.el = r.nativeElement;
-    proxyOutputs(this, this.el, ['wcsChange']);
+    proxyOutputs(this, this.el, ['wcsChange', 'wcsFocus', 'wcsBlur']);
   }
 }
 
@@ -1456,6 +1456,14 @@ export declare interface WcsSwitch extends Components.WcsSwitch {
    * Emitted when the checked property has changed.
    */
   wcsChange: EventEmitter<CustomEvent<IWcsSwitchSwitchChangeEventDetail>>;
+  /**
+   * Emitted when the switch has focus.
+   */
+  wcsFocus: EventEmitter<CustomEvent<FocusEvent>>;
+  /**
+   * Emitted when the switch loses focus.
+   */
+  wcsBlur: EventEmitter<CustomEvent<FocusEvent>>;
 }